b2marketplace

+++ database/migrations/2024_12_07_071807_create_messages_table.php
@@ -11,16 +11,19 @@ return new class extends Migration
      */
     public function up()
     {
-        Schema::create('messages', function (Blueprint $table) {
-            $table->id();
-            $table->unsignedBigInteger('sender_id');
-            $table->unsignedBigInteger('receiver_id');
-            $table->text('content');
-            $table->timestamps();
+        if (!Schema::hasTable('messages')) {
+            Schema::create('messages', function (Blueprint $table) {
+                $table->id();
+                $table->unsignedBigInteger('sender_id');
+                $table->unsignedBigInteger('receiver_id');
+                $table->text('content');
+                $table->timestamps();
+        
+                $table->foreign('sender_id')->references('id')->on('users')->onDelete('cascade');
+                $table->foreign('receiver_id')->references('id')->on('users')->onDelete('cascade');
+            });
     
-            $table->foreign('sender_id')->references('id')->on('users')->onDelete('cascade');
-            $table->foreign('receiver_id')->references('id')->on('users')->onDelete('cascade');
-        });
+        }
     }
     
     /**
